firmware: fix audio cmd stats, type mismatch, and UART spam
Three confirmed bugs in spi_protocol.c and related files:

1. BUG: audio command handlers (AUDIO_VOLUME, AUDIO_MIC_GAIN, AUDIO_PTT)
   incremented cmd_unknown_rx on short-payload or invalid-mode errors.
   cmd_unknown_rx is reserved for unrecognised opcodes (dispatch_frame
   default branch). Known commands with bad arguments should not pollute
   that counter; doing so makes it impossible to distinguish genuine
   unknown-opcode errors from framing truncation. Remove all three
   incorrect increments and add comments explaining the invariant.

2. BUG: spi_protocol_update_telemetry() declared rssi_dbm_x10,
   temp_c_x10, and cc_rssi_x10 as uint16_t even though the device_state
   fields they write are int16_t and the return types of lms7002m_read_rssi()
   and cc1101_get_rssi_x10() are int16_t. The mismatch forced callers
   to cast signed values to uint16_t at the call site (main.c), masking
   negative RSSI values. Fix: change the three parameters to int16_t in
   both the header declaration and the implementation, and remove the
   now-unnecessary casts from collect_telemetry() in main.c.

3. BUG: success-path printf() in handle_cmd_audio_volume() and
   handle_cmd_audio_mic_gain() fired on every successful command.
   These handlers can be called many times per second during active
   walkie-talkie operation (e.g., volume ramp, mic-gain adjustment).
   Per-call logging at that rate floods the UART and buries other
   diagnostic output. Remove the success-path printfs; error paths
   are still logged.

No functional changes to protocol framing, CRC validation, RF/SDR
paths, watchdog logic, or device-tree bindings.

+++ firmware/rp2350b/src/main.c
@@ -263,8 +263,8 @@ static void collect_telemetry(void)
 {
     uint16_t vbat_mv = 0;
     int16_t temp_c_x10 = 0;
-    uint16_t rssi_dbm_x10 = 0;
-    uint16_t cc_rssi_x10 = 0;
+    int16_t rssi_dbm_x10 = 0;
+    int16_t cc_rssi_x10 = 0;
     uint16_t nfc_field_mv = 0;
 
     /* Battery and temperature from ADC */
@@ -275,15 +275,14 @@ static void collect_telemetry(void)
     }
 
     /* SDR RSSI from LMS7002M — read when SDR is active.
-     * lms7002m_read_rssi() returns dBm × 10 as a signed value.
-     * Cast to uint16_t for the telemetry wire format. */
+     * lms7002m_read_rssi() returns a signed dBm × 10 value. */
     if (g_state.sdr_dma_ready) {
-        rssi_dbm_x10 = (uint16_t)lms7002m_read_rssi();
+        rssi_dbm_x10 = lms7002m_read_rssi();
     }
 
     /* CC1101 RSSI — get_rssi_x10 reads the register internally */
     if (g_state.cc1101_ready) {
-        cc_rssi_x10 = (uint16_t)cc1101_get_rssi_x10();
+        cc_rssi_x10 = cc1101_get_rssi_x10();
     }
 
     /* NFC field strength */
@@ -293,7 +292,7 @@ static void collect_telemetry(void)
 
     /* Update protocol handler telemetry cache */
     spi_protocol_update_telemetry(rssi_dbm_x10,
-                                   (uint16_t)temp_c_x10,
+                                   temp_c_x10,
                                    vbat_mv,
                                    cc_rssi_x10,
                                    nfc_field_mv);
